feat(ui): rename string file
